+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Removed `no_page_ticking` in favor of Woof's `menu_pause_demos`
   - Removed `quick_quitgame` in favor of Woof's `quit_prompt`
 - **Improved sky stretching**
+- **Made message shadows fade out when _Message Fadeout_ is enabled**
 - **Optimized initialization of _Interpolated/True-color Lighting_ and _Radial Fog_ through multithreading**
 - **Optimized initialization of generic tranmaps by caching them in files**
 - **Raised upper limits of color _Intensity_ and _Saturation_ settings to 200%**